requirements.txt
scholarly==1.7.9
psycopg2-binary==2.9.9
python-dotenv==1.0.0
beautifulsoup4==4.12.2
requests==2.31.0
SQLAlchemy==2.0.25
alembic==1.13.1
semanticscholar==0.10.0